refactor: raceResult 클래스 추가
