+++ src/main/java/domain/Cars.java
@@ -18,6 +18,7 @@ public class Cars {
     public Cars(List<Car> cars){
         this.cars = cars;
     }
+
     public List<Car> getCars(){
         return this.cars;
     }
